+++ netket/_src/ngd/sr.py
@@ -3,10 +3,12 @@ from functools import partial
 
 import jax
 import jax.numpy as jnp
-from jax.sharding import PositionalSharding
+from jax.sharding import NamedSharding
 
 from netket import jax as nkjax
 from netket.utils.types import Array
+
+from netket import config
 
 from netket._src.ngd.kwargs import ensure_accepts_kwargs
 
@@ -63,8 +65,9 @@ def _compute_sr_update(
         num_p = updates.shape[-1] // 2
         updates = updates[:num_p] + 1j * updates[num_p:]
 
-    updates = jax.lax.with_sharding_constraint(
-        updates, PositionalSharding(jax.devices()).replicate()
-    )
+    if config.netket_experimental_sharding:
+        updates = jax.lax.with_sharding_constraint(
+            updates, NamedSharding(jax.sharding.get_abstract_mesh(), jax.P())
+        )
 
     return updates, old_updates, info