+++ netket/stats/mc_stats.py
@@ -18,7 +18,7 @@ import jax
 import numpy as np
 from jax import numpy as jnp
 
-from netket.utils import config, mpi, struct
+from netket.utils import config, struct
 
 from . import mean as _mean
 from . import var as _var